@@ -18,8 +18,8 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.1.42+42
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
-  flutter: ^2.10.0
+  sdk: ">=2.12.0"
+  flutter: ^3.0.0
 
 dependencies:
 
@@ -34,13 +34,13 @@ dependencies:
   localstore: null
 
   # for making images zoomeable etc
-  photo_view: ^0.13.0
+  photo_view: ^0.14.0
 
   # for launching the platform correct map (location)
   maps_launcher: ^2.0.1
 
   # maps -> show inspection location
-  flutter_map: ^0.14.0
+  flutter_map: ^2.2.0
 
   # annotation for classes so i can stay DRY, alternatively this https://javiercbk.github.io/json_to_dart/ would also work
   json_serializable: ^6.1.1
@@ -99,6 +99,9 @@ dependencies:
   cupertino_icons: ^1.0.2
   tuple: ^2.0.0
 
+  # for automatically generating the option toggles from the option class
+  # reflectable: ^3.0.9
+
 dev_dependencies:
   flutter_test:
     sdk: flutter
@@ -111,7 +114,7 @@ dev_dependencies:
   intl_utils: ^2.6.0
 
   # generate icons for all platforms from a single png, run flutter pub run flutter_launcher_icons:main
-  flutter_launcher_icons: "^0.9.1"
+  flutter_launcher_icons: ^0.10.0
 
 flutter_icons:
   android: "launcher_icon"
