@@ -1,292 +1,178 @@
 import 'dart:async';
-import 'dart:convert';
 import 'dart:io';
 
-import 'package:MBG_Inspektionen/classes/data/checkpointdefect.dart';
-import 'package:MBG_Inspektionen/classes/imageData.dart';
-import 'package:MBG_Inspektionen/classes/requestData.dart' show RequestData;
+import 'package:MBG_Inspektionen/backend/local.dart';
+import 'package:MBG_Inspektionen/backend/remote.dart';
 import 'package:connectivity_plus/connectivity_plus.dart';
-import 'package:flutter/cupertino.dart';
-import 'package:flutter/foundation.dart';
-import 'package:flutter_dotenv/flutter_dotenv.dart';
+import 'package:flutter/material.dart';
 import 'package:http/http.dart' as http;
 import 'package:image_picker/image_picker.dart';
-import 'package:MBG_Inspektionen/assets/consts.dart';
-import 'package:MBG_Inspektionen/classes/data/checkcategory.dart';
-import 'package:MBG_Inspektionen/classes/data/checkpoint.dart';
 import 'package:MBG_Inspektionen/classes/dropdownClasses.dart';
-import 'package:tuple/tuple.dart';
+import '../classes/imageData.dart';
+import '../extension/future.dart';
 import '../generated/l10n.dart';
-import '../helpers/toast.dart';
 import '/classes/exceptions.dart';
 import '/classes/user.dart';
-import '/extension/future.dart';
+import 'package:MBG_Inspektionen/options.dart';
 
-import './offlineProvider.dart' as OP;
 import './helpers.dart' as Helper;
 
-String routesFromData<DataT extends Data>(DataT? data) =>
-    '/${Helper.getIdentifierFromData(data)}/get';
-
-const _getImageFromHash_r = '/image/get';
-const _uploadImage_r = "/image/set";
-
-const _addNew_r = "/set";
-const _update_r = "/update";
-const _delete_r = "/delete"; // issue #36
-
-const _deleteImageByHash_r = "/deleteImgH"; // issue #39
-const _setMainImageByHash_r = "/setMainImgH"; // issue #20
-
-extension _Parser on http.BaseResponse {
-  http.Response? forceRes() {
-    try {
-      return this as http.Response;
-    } catch (e) {
-      return null;
-    }
-  }
-}
-
 /// backend Singleton to provide all functionality related to the backend
-class Backend {
+class API {
+  final remote = Remote();
+  final local = LocalMirror();
   // MARK: internals
 
-  static final Backend _instance = Backend._internal();
-  factory Backend() => _instance;
+  static final API _instance = API._internal();
+  factory API() => _instance;
 
-  Backend._internal() {
+  API._internal() {
     // init
   }
-
-  final _baseurl = dotenv.env['API_URL'];
-  final _api_key = dotenv.env['API_KEY'] ?? "apitestkey";
 
   User? _user;
 
   /// returns the currently logged in [User], whether its already initialized or not.
   /// should be prefered over [_user], since it makes sure to have it initialized
+  // ignore: non_constant_identifier_names
   Future<User?> get _c_user async {
     if (_user != null) return _user;
-    return await User.fromStore();
+    final user = await User.fromStore();
+    _user = user;
+    remote.injectUser(user);
+    return _user;
+  }
+
+  bool? _dataPrefersCache(Data? data,
+      {required Helper.SimulatedRequestType type}) {
+    bool? itPrefersCache;
+    if (type == Helper.SimulatedRequestType.GET ||
+        Options().tryOnlineUploadRequestsInCachedMode)
+      try {
+        itPrefersCache = (data as WithOffline).forceOffline;
+      } catch (e) {}
+    return itPrefersCache;
+  }
+
+  Stream<T> _run<R extends http.BaseResponse, T>({
+    required FutureOr<T> Function() offline,
+    required FutureOr<RequestAndParser<R, T>> Function() online,
+
+    /// if given and [Options] is set accordingly this is called to merge offline with online data
+    FutureOr<T> Function(T, T)? merge,
+
+    /// a callback that is called if the online request fails, it gets passed the offline result as well as the request and the parser for online result
+    ///
+    /// e.g.:
+    /// ```dart
+    /// onlineFailedCB(T onlineResult, RequestAndParser<R, T> requestAndParser) async {
+    ///   modifyReq(requestAndParser.rd);
+    /// };
+    /// ```
+    FutureOr<T> Function(T, RequestAndParser<R, T>)? onlineFailedCB,
+    required Helper.SimulatedRequestType requestType,
+    bool? itPrefersCache = false,
+  }) {
+    var controller = StreamController<T>();
+    // late var canBeClosed; // = Future.delayed(Duration(seconds: 10), () => true);
+    final _itPrefersCache = itPrefersCache ?? false;
+    late T offlineRes;
+    T onlineRes;
+    Future<bool> doOffline({bool orDontIf = false}) async {
+      if (orDontIf) return false;
+      Future<T>(offline).then((value) {
+        offlineRes = value;
+        controller.add(offlineRes);
+        return true;
+      }, onError: (err) {
+        debugPrint('offline failed: ' + err.toString());
+        return false;
+      });
+      return false;
+    }
+
+    Future<RequestAndParser<R, T>>(online).then(
+      (rap) async {
+        // final _ukey = UniqueKey();
+        late Object _latestErr;
+        Future<bool> doOnline(
+            {bool orDontIf = false, bool? forceOnline}) async {
+          // debugPrint(_ukey.toString() + 'running online');
+          try {
+            // debugPrint(orDontIf ? 'we gonna skip' : 'online');
+            if (orDontIf) {
+              throw BackendCommunicationException(
+                  'we prefer the local variant');
+            }
+            // debugPrint(_ukey.toString() + 'network?');
+            final __x = await tryNetwork(requestType: requestType);
+            // debugPrint(_ukey.toString() + 'network worked');
+            final bool wantsmerged = merge != null &&
+                Options().canBeOffline &&
+                (Options().mergeOnlineEvenInCached ||
+                    Options().mergeOnline && !_itPrefersCache);
+            final bool wantsonline = forceOnline ??
+                //TODO: okay but this might be the wrong type
+                (Options().preferRemoteData || Options().preferRemoteImgs);
+            if (wantsonline || wantsmerged) {
+              final res = await remote.postJSON(rap.rd);
+              onlineRes = await rap.parser(res as R);
+              if (wantsonline) controller.add(onlineRes);
+              if (wantsmerged)
+                controller.add(await merge(offlineRes, onlineRes));
+              // debugPrint(_ukey.toString() +
+              //     'online succeeded: ' +
+              //     onlineRes.toString());
+            }
+            // debugPrint(_ukey.toString() +
+            //     'ran online ' +
+            //     wantsonline.toString() +
+            //     wantsmerged.toString());
+          } catch (e) {
+            // debugPrint(_ukey.toString() + 'online failed: ' + e.toString());
+            _latestErr = e;
+            return false;
+          }
+          return true;
+        }
+
+        onlineFailedProcedure() async {
+          bool log = rap.rd.logIfFailed ??
+              (requestType != Helper.SimulatedRequestType.GET);
+          if (onlineFailedCB != null)
+            await onlineFailedCB(offlineRes, rap);
+          else if (log) await local.logFailedReq(rap.rd);
+          if (Options().debugLocalMirror)
+            debugPrint(
+                'failed request ${log ? "and logged it" : ""}: ${rap.rd.json} \n\t error was $_latestErr');
+        }
+
+        List<bool> _success = await Future.wait([
+          doOnline(
+            orDontIf: _itPrefersCache && !Options().mergeOnlineEvenInCached,
+          ),
+          doOffline(
+            orDontIf: !Options().canBeOffline,
+          )
+        ], eagerError: false);
+
+        bool onlineSucc = _success[0];
+        bool offlineSucc = _success[1];
+        var x = 0;
+        if (!onlineSucc && !offlineSucc && Options().tryOnlineIfOfflineFailed) {
+          // debugPrint(_ukey.toString() + 'trying online again');
+          onlineSucc = await doOnline(
+              forceOnline:
+                  true); //TODO: offline has to succeed normally so we dont fetch online if we dont *really* need to
+        }
+        controller.close();
+        if (!onlineSucc) onlineFailedProcedure();
+      },
+    );
+    return controller.stream;
   }
 
   // MARK: available Helpers
-
-  /// checks whether a connection to the backend is possible
-  /// throws [NoConnectionToBackendException] or [SocketException] if its not.
-  Future connectionGuard({
-    Duration? timeout,
-  }) async {
-    if (_baseurl == null)
-      throw NoConnectionToBackendException(
-          S.current.exceptionNoUrlToConnectToProvided);
-
-    //check network
-    var connection = await (Connectivity().checkConnectivity());
-    if (connection == ConnectivityResult.none)
-      throw NoConnectionToBackendException(S.current.noNetworkAvailable);
-    if (!Options.canUseMobileNetworkIfPossible &&
-        connection == ConnectivityResult.mobile)
-      throw NoConnectionToBackendException(S.current.mobileNetworkNotAllowed);
-
-    try {
-      // check if we can reach our api
-      await post_JSON(
-          RequestData('/login', timeout: timeout, logIfFailed: false));
-    } catch (e) {
-      throw NoConnectionToBackendException(
-          S.current.couldntReach + " $_baseurl");
-    }
-  }
-
-  /// make an actual API request to a route, and always append the API_KEY as authorization-header
-  Future<http.Response?> post(
-    String route, {
-    Map<String, String>? headers,
-    String? body,
-    Encoding? encoding,
-    Duration? timeout,
-  }) =>
-      send(
-        makepost(route, headers: headers, body: body, encoding: encoding),
-        timeout: timeout,
-      );
-
-  Future<http.Response?> send(
-    http.Request request, {
-    Duration? timeout,
-    bool returnsBinary = false,
-  }) async {
-    final req = request.send();
-    final res = (timeout == null) ? await req : await req.timeout(timeout);
-
-    final ret = await http.Response.fromStream(res); // res.forceRes();
-    if (Options.debugAllResponses && !returnsBinary)
-      debugPrint("res: " + ret.body);
-    return ret;
-  }
-
-  /// make an actual API request to a route, and always append the API_KEY as authorization-header
-  http.Request makepost(
-    String route, {
-    Map<String, String>? headers,
-    String? body,
-    Encoding? encoding,
-  }) {
-    headers = headers ?? {};
-    headers.addAll({HttpHeaders.authorizationHeader: _api_key});
-    var fullURL = Uri.parse(_baseurl! + route);
-    final req = http.Request('post', fullURL)..headers.addAll(headers);
-    if (encoding != null) req.encoding = encoding;
-    if (body != null) req.body = body;
-    return req;
-  }
-
-  /// post_JSON to our backend as the user
-  Future<http.BaseResponse?> post_JSON(RequestData rd) async {
-    var headers = {HttpHeaders.contentTypeHeader: 'application/json'};
-    rd.json = rd.json ?? {};
-    rd.json!['user'] = (await _c_user)?.toJson();
-    if (Options.debugAllResponses) debugPrint('req: ' + jsonEncode(json));
-    try {
-      if (rd.multipart_files.isNotEmpty) {
-        http.MultipartRequest? mreq;
-        try {
-          var fullURL = Uri.parse(_baseurl! + rd.route);
-          mreq = http.MultipartRequest('POST', fullURL)
-            ..files.addAll(
-              List<http.MultipartFile>.from((await Future.wait(
-                rd.multipart_files.map(
-                  (xfile) => http.MultipartFile.fromPath('package', xfile.path,
-                      filename: xfile.name),
-                ),
-              ))
-                  .whereType<http.MultipartFile>()),
-            )
-            ..headers.addAll({HttpHeaders.authorizationHeader: _api_key})
-            ..fields.addAll(/*flatten()*/ rd.json!.map<String, String>(
-                (key, value) => MapEntry(key, value.toString())));
-          debugPrint("gonna send multipart-req with booty ${mreq.fields}");
-          var res = (rd.timeout == null)
-              ? await mreq
-                  .send() //XXX: use client.send(mreq) for every request here
-              : await mreq.send().timeout(rd.timeout!);
-          return res;
-        } on Exception catch (e) {
-          if (rd.logIfFailed) OP.logFailedReq(rd);
-          debugPrint('multipartRequest, failed');
-          throw e;
-        }
-      } else {
-        final req =
-            makepost(rd.route, headers: headers, body: jsonEncode(rd.json));
-        try {
-          return await send(
-            req,
-            timeout: rd.timeout,
-            returnsBinary: rd.returnsBinary,
-          );
-        } catch (e) {
-          if (rd.logIfFailed) OP.logFailedReq(rd);
-          debugPrint('request failed');
-          throw e;
-        }
-      }
-    } catch (e) {
-      debugPrint("request failed, cause : ${e}");
-      return null;
-    }
-  }
-
-  //final _imageStreamController = BehaviorSubject<String>();
-  Stream<ImageData?> _fetchImage(String hash) async* {
-    bool cacheHit = false;
-    if (Options.canBeOffline)
-      try {
-        final img = await OP.readImage(hash);
-        if (img == null) throw Exception("no img cached");
-        yield ImageData(img, id: hash);
-        cacheHit = true;
-      } catch (e) {
-        //yield null;
-      }
-    if (!cacheHit || Options.preferRemoteImages) {
-      http.Response? res = (await post_JSON(RequestData(
-        _getImageFromHash_r,
-        json: {
-          'imghash': hash,
-        },
-        returnsBinary: true,
-        logIfFailed: false,
-      )))
-          ?.forceRes();
-      if (res == null || res.statusCode != 200)
-        yield null;
-      else {
-        try {
-          await OP.storeImage(res.bodyBytes, hash);
-          yield ImageData(
-            (await OP.readImage(hash))!,
-            id: hash,
-          );
-        } catch (e) {
-          debugPrint("failed to load webimg: " + e.toString());
-        }
-      }
-    }
-  }
-
-  //final _imageStreamController = BehaviorSubject<String>();
-  Future<ImageData?> _fetchImage_fut(String hash) async {
-    bool cacheHit = false;
-    if (Options.canBeOffline && !Options.preferRemoteImages) {
-      try {
-        final img = await OP.readImage(hash);
-        if (img == null) throw Exception("no img cached");
-        cacheHit = true;
-        return ImageData(img, id: hash);
-      } catch (e) {
-        //yield null;
-      }
-    }
-    if (!cacheHit || Options.preferRemoteImages) {
-      http.Response? res = (await post_JSON(RequestData(
-        _getImageFromHash_r,
-        json: {
-          'imghash': hash,
-        },
-        returnsBinary: true,
-        logIfFailed: false,
-      )))
-          ?.forceRes();
-      if (res == null || res.statusCode != 200)
-        return null;
-      else {
-        try {
-          await OP.storeImage(res.bodyBytes, hash);
-          return ImageData(
-            (await OP.readImage(hash))!,
-            id: hash,
-          );
-        } catch (e) {
-          debugPrint("failed to load webimg: " + e.toString());
-        }
-      }
-    }
-  }
-
-  Future<DataT?> Function(Map<String, dynamic>)
-      _generateImageFetcher<DataT extends Data>(
-    DataT? Function(Map<String, dynamic>) jsoner,
-  ) {
-    // only fetch first image automagically and the others only when said so (or at least not make the UI wait for it (#34, #35))
-    return (Map<String, dynamic> json) async {
-      DataT? data = jsoner(json);
-      if (data == null) return null;
-      if (data.imagehashes == null ||
-          data.imagehashes!.length == 0) //the second check *could* be omitted
-        return data;
 
       int first_working_image_index = 0;
       String __hash = data.imagehashes![first_working_image_index];
@@ -319,132 +205,29 @@ class Backend {
           .toList()
           .sublist((__hash == Options.no_image_placeholder_name) ? 1 : 0);
 
-      return data;
-    };
+  /// checks whether a connection to the network is allowed
+  /// throws [NoConnectionToBackendException] or [SocketException] if its not.
+  Future tryNetwork({
+    Duration? timeout,
+    required Helper.SimulatedRequestType requestType,
+  }) async {
+    // debugPrint('network testing..');
+    //check network
+    if (Options().forceOffline)
+      throw NoConnectionToBackendException(
+          S.current.nonetwork_forcedOfflineMode);
+    final connection = await (Connectivity().checkConnectivity());
+    if (connection == ConnectivityResult.none)
+      throw NoConnectionToBackendException(S.current.noNetworkAvailable);
+    if (connection == ConnectivityResult.mobile &&
+        ((requestType != Helper.SimulatedRequestType.GET &&
+                !Options().useMobileNetworkForUpload) ||
+            !Options().useMobileNetworkForDownload))
+      throw NoConnectionToBackendException(S.current.mobileNetworkNotAllowed);
+    // debugPrint('network good');
   }
 
   Future<String> get rootID async => (await user)!.name;
-
-  /// Helper function to get the next [Data] (e.g. all [CheckPoint]s for chosen [CheckCategory])
-  Stream<List<ChildData>>
-      _getAllForNextLevel<ChildData extends Data, ParentData extends Data>({
-    required String route,
-    required String jsonResponseID,
-    Map<String, dynamic>? json,
-    required ChildData? Function(Map<String, dynamic>) fromJson,
-    String? id,
-    forceOffline = false,
-  }) async* {
-    bool _forceOffline = forceOffline ?? false;
-    // yield [];
-    assert(
-        (await user) != null, S.current.wontFetchAnythingSinceNoOneIsLoggedIn);
-    String _id = id ?? json?['local_id'] ?? await rootID;
-    Map<String, dynamic> _json = {};
-    Future<ChildData?> Function(Map<String, dynamic>) imageFetcher =
-        _generateImageFetcher(fromJson);
-
-    Future<List<ChildData>> __parse(__json) => getListFromJson(
-          __json,
-          imageFetcher,
-          // _generateImageFetcher(fromJson),
-          objName: jsonResponseID,
-        );
-
-    List<ChildData>? cached;
-    if (Options.canBeOffline)
-      try {
-        _json = jsonDecode("{\"$jsonResponseID\": ${jsonEncode(
-            // Ähhh ja die liste muss wie die response aussehen damit die function weiter unten die images fetchet
-            (await OP.getAllChildrenFrom<ChildData>(_id)))}}");
-        cached = await __parse(_json);
-        yield cached;
-      } catch (e) {
-        debugPrint("couldnt read data from disk..: " + e.toString());
-      }
-
-    if (!_forceOffline ||
-        Options.mergeLoadedDataIntoOnlineDataEvenInCachedParent)
-      try {
-        final res = (await post_JSON(RequestData(
-          route,
-          json: json,
-          timeout: Duration(seconds: 10),
-          logIfFailed: false,
-        )));
-
-        final body = res!.forceRes()!.body;
-        _json = jsonDecode(body);
-        var datapoints = await __parse(_json);
-        yield datapoints;
-        if ((Options.mergeLoadedDataIntoOnlineData ||
-                Options.mergeLoadedDataIntoOnlineDataEvenInCachedParent) &&
-            cached != null) {
-          try {
-            cached.retainWhere(
-                (element) => (element as WithOffline).forceOffline);
-            var cachedIds = cached.map((element) => element.id).toList();
-            datapoints.retainWhere((element) =>
-                element.id != null && !cachedIds.contains(element.id));
-            datapoints.addAll(cached);
-            yield datapoints;
-          } catch (e) {
-            debugPrint("error merging data: " + e.toString());
-          }
-        }
-        for (var data in datapoints) {
-          String childId = await OP.storeData(data, forId: _id);
-          if (Options.debugLocalMirror)
-            debugPrint("stored new child with id: " + childId);
-        }
-      } catch (e) {
-        debugPrint("couldnt reach API: " + e.toString());
-      }
-
-    if (Options.debugAllResponses)
-      debugPrint("_getAllForNextLevel received: " + jsonEncode(json));
-  }
-
-  /// sends a [DataT] with the corresponding identifier to the given route
-  Future<http.Response?> _sendDataToRoute<DataT extends Data>(
-      {required DataT? data,
-      required String route,
-      Map<String, dynamic> other = const {},
-      bool networkIsCrucial = false}) async {
-    if (networkIsCrucial || !Options.canBeOffline)
-      try {
-        await connectionGuard();
-      } catch (error) {
-        showToast(error.toString() + "\n" + S.current.tryAgainLater_noNetwork);
-        return null;
-      }
-
-    if (Options.debugAllResponses)
-      debugPrint(
-          "we send this data to ${route}:" + (data?.toJson().toString() ?? ""));
-    if (data == null) return null;
-    var json_data = data.toJson();
-    http.Response? res = (await post_JSON(RequestData(route, json: {
-      'type': Helper.getIdentifierFromData(data),
-      'data': json_data,
-      ...other
-    })))
-        ?.forceRes();
-    if (Options.debugAllResponses)
-      debugPrint("and we received :" + (res?.body.toString() ?? ""));
-
-    if (res?.statusCode != 200) {
-      _maybeShowToast(
-          "${S.current.anUnknownErrorOccured}, ${res?.statusCode}: ${res?.reasonPhrase}");
-    }
-    return res;
-  }
-
-  void _maybeShowToast(String? message) {
-    if (message != null && message != "") {
-      showToast(message);
-    }
-  }
 
   // MARK: API
 
@@ -459,21 +242,18 @@ class Backend {
 
   /// login a [User] by checking if he exists in the remote database
   Future<DisplayUser?> login(User user) async {
+    // final requestType =
+    //     Helper.SimulatedRequestType.GET;
     // if user is already logged in
-    if (await isUserLoggedIn(user)) return await this.user;
-    await connectionGuard();
-    _user = user;
-    var res = (await post_JSON(RequestData('/login', logIfFailed: false)))
-        ?.forceRes();
-    if (res != null && res.statusCode == 200) {
-      //success
-      var resb = jsonDecode(res.body)['user'];
-      _user?.fromMap(resb);
-      await _user?.store();
+    if (await isUserLoggedIn(user)) return this.user;
+    try {
+      final _user = await remote.login(user);
+      await _user!.store();
       return this.user;
+    } catch (e) {
+      logout();
+      rethrow;
     }
-    await logout(); //we could omit the await for a slight speed improvement (but if anything crashes for some reason it could lead to unexpected behaviour)
-    throw ResponseException(res);
   }
 
   /// removes the credentials from local storage and therefors logs out
@@ -489,26 +269,30 @@ class Backend {
       getNextDatapoint<ChildData extends Data, ParentData extends WithOffline?>(
     ParentData data,
   ) async* {
-    if (!Options.canBeOffline)
-      try {
-        await connectionGuard();
-      } catch (error) {
-        showToast(error.toString() + "\n" + S.current.tryAgainLater_noNetwork);
-        yield [];
-      }
-    final childTypeStr = Helper.getIdentifierFromData<ChildData>(null);
-    if (childTypeStr == null) throw Exception('type not supported');
-    await for (var l in _getAllForNextLevel(
-      route: routesFromData<ChildData>(null),
-      jsonResponseID: childTypeStr + 's',
-      json: data?.toSmallJson(),
-      fromJson: (json) => /*Child*/ Data.fromJson<ChildData>(json),
-      id: data?.id,
-      forceOffline: data?.forceOffline,
-    )) {
-      // debugPrint('new value $l');
-      yield l;
-    }
+    final requestType = Helper.SimulatedRequestType.GET;
+    assert((await API().user) != null,
+        S.current.wontFetchAnythingSinceNoOneIsLoggedIn);
+
+    yield* _run(
+      itPrefersCache: _dataPrefersCache(data, type: requestType),
+      offline: () => local.getNextDatapoint(data),
+      online: () => remote.getNextDatapoint(data),
+      requestType: requestType,
+      merge: (cached, upstream) async {
+        try {
+          cached
+              .retainWhere((element) => (element as WithOffline).forceOffline);
+          var cachedIds = cached.map((element) => element.id).toList();
+          upstream.retainWhere((element) =>
+              element.id != null && !cachedIds.contains(element.id));
+          upstream.addAll(cached);
+          return upstream;
+        } catch (e) {
+          debugPrint("error merging data: " + e.toString());
+          return cached;
+        }
+      },
+    );
   }
 
   /// sets a new [DataT]
@@ -516,17 +300,14 @@ class Backend {
     DataT? data, {
     Data? caller,
   }) async {
-    //offline procedure, needs some stuff changed and added..
-    if (caller != null && data != null && caller.id != null) {
-      data.id = /*'_on_' + */ (data.id ?? '__new__' + data.title);
-      OP.storeData<DataT>(data, forId: caller.id!);
-    }
+    final requestType = Helper.SimulatedRequestType.PUT;
 
-    var body = (await _sendDataToRoute(
-            data: data, route: _addNew_r, networkIsCrucial: true))
-        ?.body;
-    //XXX if the resulting Data is needed we would need to pass it correctly from this response body, the following just returns the input on success
-    return body == null ? null : data;
+    return _run(
+      itPrefersCache: _dataPrefersCache(caller, type: requestType),
+      offline: () => local.setNew(data, caller: caller),
+      online: () => remote.setNew(data),
+      requestType: requestType,
+    ).last;
   }
 
   /// updates a [DataT] and returns the response
@@ -535,15 +316,13 @@ class Backend {
     Data? caller,
     bool forceUpdate = false,
   }) async {
-    //offline procedure, needs some stuff changed and added..
-    if ((forceUpdate || caller != null && caller.id != null) && data != null) {
-      data.id = /*'_oe_' + */ (data.id ?? '__new__' + data.title);
-      OP.storeData<DataT>(data, forId: caller?.id ?? await rootID);
-    }
-
-    return (await _sendDataToRoute(
-            data: data, route: _update_r, networkIsCrucial: true))
-        ?.body;
+    final requestType = Helper.SimulatedRequestType.PUT;
+    return _run(
+      itPrefersCache: _dataPrefersCache(caller, type: requestType),
+      offline: () => local.update(data, caller: caller),
+      online: () => remote.update(data),
+      requestType: requestType,
+    ).last;
   }
 
   /// deletes a [DataT] and returns the response
@@ -551,26 +330,39 @@ class Backend {
     DataT? data, {
     Data? caller,
   }) async {
-    //offline procedure, needs some stuff changed and added..
-    if (caller != null &&
-        data != null &&
-        caller.id != null &&
-        data.id != null) {
-      OP.deleteData<DataT>(data.id!, parentId: caller.id!);
-    }
-
-    return (await _sendDataToRoute(
-            data: data, route: _delete_r, networkIsCrucial: true))
-        ?.body;
+    final requestType = Helper.SimulatedRequestType.DELETE;
+    return _run(
+      itPrefersCache: _dataPrefersCache(caller, type: requestType),
+      offline: () => local.delete(data, caller: caller),
+      online: () => remote.delete(data),
+      requestType: requestType,
+    ).last;
   }
 
   /// deletes an image specified by its hash and returns the response
-  Future<String?> deleteImageByHash(String hash) async {
-    await OP.deleteImage(hash);
-    return (await post_JSON(
-            RequestData(_deleteImageByHash_r, json: {'hash': hash})))
-        ?.forceRes()
-        ?.body;
+  Future<ImageData?> getImageByHash(String hash) async {
+    final requestType = Helper.SimulatedRequestType.GET;
+
+    return _run(
+      itPrefersCache: !Options().preferRemoteImgs,
+      offline: () => local.getImageByHash(hash),
+      online: () => remote.getImageByHash(hash),
+      requestType: requestType,
+    ).last;
+  }
+
+  /// deletes an image specified by its hash and returns the response
+  Future<String?> deleteImageByHash<DataT extends Data>(
+    DataT? data,
+    String hash,
+  ) async {
+    final requestType = Helper.SimulatedRequestType.DELETE;
+    return _run(
+      itPrefersCache: _dataPrefersCache(data, type: requestType),
+      offline: () => local.deleteImageByHash(hash),
+      online: () => remote.deleteImageByHash(hash),
+      requestType: requestType,
+    ).last;
   }
 
   /// sets an image specified by its hash as the new main image
@@ -580,152 +372,63 @@ class Backend {
     Data? caller,
     bool forceUpdate = false,
   }) async {
-    //offline procedure, needs some stuff changed and added..
-    if ((forceUpdate || caller != null && caller.id != null) && data != null) {
-      try {
-        data.id = /*'_oe_' + */ (data.id ?? '__new__' + data.title);
-        data.imagehashes!.remove(hash);
-        data.imagehashes!.insert(0, hash);
-        OP.storeData<DataT>(data, forId: caller?.id ?? await rootID);
-      } catch (e) {
-        debugPrint('failed to update main image locally');
-      }
-    }
-    return (await _sendDataToRoute(
-      data: data,
-      route: _setMainImageByHash_r,
-      other: {
-        'hash': hash,
-      },
-    ))
-        ?.body;
+    final requestType = Helper.SimulatedRequestType.PUT;
+    return _run(
+      itPrefersCache: _dataPrefersCache(data, type: requestType),
+      offline: () => local.setMainImageByHash(
+        data,
+        hash,
+        caller: caller,
+        forceUpdate: forceUpdate,
+      ),
+      online: () => remote.setMainImageByHash(
+        data,
+        hash,
+      ),
+      requestType: requestType,
+    ).last;
   }
 
   /// upload a bunch of images
   Future<String?> uploadFiles<DataT extends Data>(
     DataT data,
-    List<XFile> files,
-  ) async {
-    ////ODO: we currently store everything n the root dir, but we want to add into specific subdir that needs to be extracted from rew.body.E1 etc
-    debugPrint('uploading images ${files}');
-    var json_data = data.toJson();
-    var res = await post_JSON(RequestData(
-      _uploadImage_r,
-      json: {
-        'type': Helper.getIdentifierFromData(data),
-        'data': json.encode(json_data),
-      },
-      multipart_files: files,
-    )); //wont work
-    if (res?.statusCode != 200) {
-      debugPrint('not ok: ${res?.statusCode.toString()}');
-      // debugPrint(res?.contentLength.toString());
-    }
-    return (res.runtimeType == http.Response)
-        ? (res as http.Response?)?.body //TODO meh remove crash und stuff
-        : await (res as http.StreamedResponse?)?.stream.bytesToString();
-  }
-
-  /// removes all locally stored images via [OP]
-  final deleteCache = OP.deleteAll;
-
-  Future<bool> retryFailedrequests() async {
-    final failedReqs = await OP.getAllFailedRequests() ?? [];
-    bool success = true;
-    for (final reqd in failedReqs) {
-      final docID = reqd.item1;
-      final rd = reqd.item2;
-      if (rd != null)
-        try {
-          final res = await Backend().post_JSON(rd);
-          //nur 200er als ok einstufen
-          if (res!.statusCode == 200) {
-            OP.failedRequestWasSuccesful(docID);
-          } else {
-            success = false;
-            //TODO: what todo here?
-            break;
-          }
-        } catch (e) {
-          debugPrint('failed to retry request: $e');
-          success = false;
-        }
-    }
-    return success;
-  }
-
-  //TODO: das klappt zwar, aber das abspeichern selbst oder anzeigen nicht, wird aber OP liegen
-  /// recursivle cache all elements that underly the caller
-  Future<bool> loadAndCacheAll<
-      ChildData extends WithLangText,
-      ParentData extends WithOffline,
-      DDModel extends DropDownModel<ChildData, ParentData>>(
-    DDModel caller,
-    int depth, {
-    String? name,
-    String? parentID,
+    List<XFile> files, {
+    Data? caller,
+    bool forceUpdate = false,
   }) async {
-    //base-case: CheckPointDefects have no children
-    // if (typeOf<ChildData>() == CheckPointDefect) return true;//XXX: shit, this generic bums wont work
-    if (depth == 0) return true;
-    depth--;
-    try {
-      //fail early if no connection
-      await connectionGuard();
-      //get all children, this will also cache them internally
-      var children = await caller.all.last;
-      var didSucceed = await Future.wait(children.map((child) async {
-        String _name = '$name -> ${child.title}';
-        debugPrint('__1234 got $depth: $_name');
-        if (depth == 0)
-          return true; //base-case as to not call generateNextModel
-        bool child_succeeded = await loadAndCacheAll(
-            caller.generateNextModel(child), depth,
-            name: _name, parentID: caller.currentData.id);
-        // try {
-
-        // } catch (e) {}
-        return child_succeeded;
-      }));
-
-      //if all children succeeded recursive calling succeeded
-      bool success = didSucceed.every((el) => el);
-      if (success) {
-        caller.currentData.forceOffline = true;
-        if (parentID == null) return false;
-
-        try {
-          var id = await OP.storeData(caller.currentData, forId: parentID);
-          debugPrint(
-              '_32 loading. ${depth + 1} stored : ${id}  ${caller.currentData.title}');
-        } catch (e) {
-          return false;
-        }
-      }
-
-      return success;
-    } catch (error) {
-      debugPrint('failed! ${depth + 1}');
-      showToast(error.toString() + "\n" + S.current.tryAgainLater_noNetwork);
-      return false; //failed
-    }
+    final requestType = Helper.SimulatedRequestType.PUT;
+    return _run(
+      itPrefersCache: _dataPrefersCache(data, type: requestType),
+      offline: () => local.uploadFiles(
+        data,
+        files,
+        caller: caller,
+        forceUpdate: forceUpdate,
+      ),
+      online: () => remote.uploadFiles(
+        data,
+        files,
+      ),
+      onlineFailedCB: (onlineRes, rap) {},
+      requestType: requestType,
+    ).last;
   }
 }
 
-/// Helper function to parse a [List] of [Data] Objects from a Json-[Map]
-Future<List<T>> getListFromJson<T extends Data>(Map<String, dynamic> json,
-    FutureOr<T?> Function(Map<String, dynamic>) converter,
-    {String? objName}) async {
-  try {
-    List<dynamic> str = (objName != null) ? json[objName] : json;
-    return List<T>.from(
-        (await Future.wait(str.map((elem) async => await converter(elem))))
-            .whereType<T>());
-  } catch (e) {
-    debugPrint(
-        'could not parse response: ' + e.toString() + '<--' + jsonEncode(json));
-    throw BackendCommunicationException(
-        S.current.couldNotParseResponse + jsonEncode(json));
-  }
-  //return [];
+D injectImages<D extends WithImgHashes>(D data) {
+  if (data.imagehashes == null ||
+      data.imagehashes!.length == 0) //the second check *could* be omitted
+    return data;
+  String _firstHash = data.imagehashes![0];
+  data.mainImage = (_firstHash == Options().no_image_placeholder_name)
+      ? null
+      : API().getImageByHash(_firstHash);
+  data.imageFutures = data.imagehashes
+      ?.map((hash) => API().getImageByHash(hash))
+      .toList()
+      .sublist((_firstHash == Options().no_image_placeholder_name) ? 1 : 0);
+  data.previewImage = data.imageFutures != null
+      ? data.imageFutures!.ordered_firstNonNull
+      : Future.value(null);
+  return data;
 }
