@@ -17,13 +17,13 @@ InspectionLocation _$InspectionLocationFromJson(Map<String, dynamic> json) =>
       stONr: json['StONr'] as int,
       strasse: json['Straße'] as String?,
       fallback_coords: _toplevelhelperLatLng_fromJson(
-          json['latLng'] as Map<String, dynamic>),
+          json['latLng'] as Map<String, dynamic>?),
     )
       ..langText = json['langText'] as String?
       ..forceOffline_nullable = json['offline'] as bool?
       ..id = json['local_id'] as String?
       ..eigentuemer = json['Eigentuemer'] as String?
-      ..bauwerkhoehe = json['Bauwerkhoehe'] as double?
+      ..bauwerkhoehe = (json['Bauwerkhoehe'] as num?)?.toDouble()
       ..baujahr = json['Baujahr'] as int?
       ..ansprechpartner = json['Ansprechpartner'] as String?
       ..steigwegtyp = json['Steigwegtyp'] as String?
