@@ -1,4 +1,5 @@
-import 'package:flutter/material.dart';
+// ignore_for_file: non_constant_identifier_names
+
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
