@@ -1,14 +1,11 @@
-import 'package:MBG_Inspektionen/assets/consts.dart';
-import 'package:MBG_Inspektionen/backend/api.dart';
+import 'package:MBG_Inspektionen/options.dart';
 import 'package:flutter/material.dart';
-import 'package:MBG_Inspektionen/pages/login/loginModel.dart';
 import 'package:MBG_Inspektionen/pages/settings/settingsView.dart';
-import 'package:MBG_Inspektionen/widgets/MyListTile1.dart';
 import 'package:MBG_Inspektionen/fragments/RandomDogsScrollView.dart';
 import 'package:package_info_plus/package_info_plus.dart';
-import 'package:provider/provider.dart';
 
 import '../generated/l10n.dart';
+import '../widgets/openNewViewTile.dart';
 
 class MainDrawer extends StatelessWidget {
   @override
@@ -35,9 +32,9 @@ class MainDrawer extends StatelessWidget {
         children: <Widget>[
           MainDrawerHeader(),
           settingsTile,
-          if (Options.showDoggo) unlimiteddoggosTile,
+          if (Options().showDoggo) unlimiteddoggosTile,
           Spacer(),
-          if (Options.canBeOffline) UploadSyncTile(),
+          if (Options().canBeOffline) UploadSyncTile(),
           SafeArea(
             child: Padding(
               padding: const EdgeInsets.all(8.0),
@@ -66,43 +63,6 @@ class VersionInfo extends StatelessWidget {
                 (snapshot.data?.buildNumber ?? '...'),
           );
         });
-  }
-}
-
-class OpenNewViewTile extends StatelessWidget {
-  final IconData icon;
-  final String title;
-  final Widget newView;
-  final bool supplyLoginModel;
-
-  const OpenNewViewTile({
-    this.icon = Icons.open_in_full,
-    required this.title,
-    required this.newView,
-    this.supplyLoginModel = false,
-    Key? key,
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return MyCardListTile1(
-      icon: icon,
-      text: title,
-      onTap: () {
-        Navigator.pop(context);
-        Navigator.push(
-          context,
-          MaterialPageRoute(
-            builder: (newcontext) => supplyLoginModel
-                ? ChangeNotifierProvider<LoginModel>.value(
-                    value: Provider.of<LoginModel>(context),
-                    child: newView,
-                  )
-                : newView,
-          ),
-        );
-      },
-    );
   }
 }
 
