@@ -1,3 +1,5 @@
+// ignore_for_file: unused_element
+
 import 'package:MBG_Inspektionen/classes/imageData.dart';
 import 'package:MBG_Inspektionen/generated/l10n.dart';
 import 'package:flutter/material.dart';
@@ -47,8 +49,10 @@ class ExpandableCard2 extends ExpandablesRadio {
   }
 
   Color _color(isExpanded, context) => !isExpanded
-      ? (Theme.of(context).iconTheme.color ?? Theme.of(context).disabledColor)
-      : Theme.of(context).colorScheme.secondary;
+      // ? (Theme.of(context).iconTheme.color ?? Theme.of(context).disabledColor)
+      // : Theme.of(context).colorScheme.secondary;
+      ? Theme.of(context).disabledColor
+      : Theme.of(context).iconTheme.color!;
 
   @override
   ExpansionPanelRadio make(BuildContext context) {
@@ -84,9 +88,10 @@ class ExpandableCard2 extends ExpandablesRadio {
               title: Text(
                 title + (isExpanded ? ':' : ''),
                 style: isExpanded
-                    ? Theme.of(context).textTheme.headline5?.apply(
-                          color: Theme.of(context).colorScheme.primary,
-                        )
+                    ? Theme.of(context).textTheme.headline5
+                    // ?.apply(
+                    //       color: Theme.of(context).colorScheme.primary,
+                    //     )
                     : Theme.of(context).textTheme.bodyText1,
               ),
               subtitle: (subtitle != null) ? Text(subtitle ?? "") : null,
